Move telepath module to components
